% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Dispersion_analysis.R
\name{mod_Dispersion_analysis_ui}
\alias{mod_Dispersion_analysis_ui}
\title{Dispersion_analysis UI Function}
\usage{
mod_Dispersion_analysis_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module runs Dispersions analysis
}
\author{
Audrey BEAUFILS
}
