% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GSEA.R
\name{mod_GSEA_ui}
\alias{mod_GSEA_ui}
\title{GSEA UI Function}
\usage{
mod_GSEA_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module run Gen Set Enrichment Analysis on GO terms.
}
\author{
Audrey BEAUFILS
}
