% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GSEAKEGG.R
\name{mod_GSEAKEGG_ui}
\alias{mod_GSEAKEGG_ui}
\title{GSEAKEGG UI Function}
\usage{
mod_GSEAKEGG_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module run Gene Set Enrichment Analysis on Kegg pathway.
}
\author{
Audrey BEAUFILS
}
