% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GSEAReactome.R
\name{mod_GSEAReactome_ui}
\alias{mod_GSEAReactome_ui}
\title{GSEAReactome UI Function}
\usage{
mod_GSEAReactome_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module runs Over Representation Analysis on Reactome pathway.
}
\author{
Audrey BEAUFILS
}
