% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home.R
\name{mod_home_ui}
\alias{mod_home_ui}
\title{home UI Function}
\usage{
mod_home_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module displays the home page
}
\author{
Audrey BEAUFILS
}
