% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Input_Norm.R
\name{mod_Input_Norm_ui}
\alias{mod_Input_Norm_ui}
\title{Input_Norm UI Function}
\usage{
mod_Input_Norm_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module displays parameter for differential Analysis
}
\author{
Audrey BEAUFILS
}
