% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Normalisation_AD.R
\name{mod_Normalisation_AD_ui}
\alias{mod_Normalisation_AD_ui}
\title{Normalisation_AD UI Function}
\usage{
mod_Normalisation_AD_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module normalize counts
}
\author{
Audrey BEAUFILS
}
