% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ORA.R
\name{mod_ORA_ui}
\alias{mod_ORA_ui}
\title{ORA UI Function}
\usage{
mod_ORA_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module runs Over Representation Analysis on GO terms.
}
\author{
Audrey BEAUFILS
}
