% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ORAKEGG.R
\name{mod_ORAKEGG_ui}
\alias{mod_ORAKEGG_ui}
\title{ORAKEGG UI Function}
\usage{
mod_ORAKEGG_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module runs Over Representation Analysis on KEGG pathway.
}
\author{
Audrey BEAUFILS
}
