% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parameter_EA.R
\name{mod_parameter_EA_ui}
\alias{mod_parameter_EA_ui}
\title{parameter_EA UI Function}
\usage{
mod_parameter_EA_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module displays parameter for Enrichment Analysis
}
\author{
Audrey BEAUFILS
}
