% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Replica_Quality.R
\name{mod_Replica_Quality_ui}
\alias{mod_Replica_Quality_ui}
\title{Replica_Quality UI Function}
\usage{
mod_Replica_Quality_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
This module run control quality on Replica.
}
\author{
Audrey BEAUFILS
}
